restructure folders for components
