+++ packages/features/insights/components/BookedByCell.tsx
@@ -2,7 +2,7 @@ import { useId } from "react";
 
 import { Badge } from "@calcom/ui/components/badge";
 
-import type { RoutingFormTableRow } from "../lib/types";
+import type { RoutingFormTableRow } from "../../lib/types";
 import { CellWithOverflowX } from "./CellWithOverflowX";
 
 export function BookedByCell({
