+++ packages/features/insights/components/BookingAtCell.tsx
@@ -13,7 +13,7 @@ import {
 } from "@calcom/ui/components/hover-card";
 import { Icon } from "@calcom/ui/components/icon";
 
-import type { RoutingFormTableRow } from "../lib/types";
+import type { RoutingFormTableRow } from "../../lib/types";
 import { BookingStatusBadge } from "./BookingStatusBadge";
 
 export function BookingAtCell({ row, rowId }: { row: RoutingFormTableRow; rowId: number }) {
