+++ packages/features/insights/components/BookingStatusBadge.tsx
@@ -2,7 +2,7 @@ import { BookingStatus } from "@calcom/prisma/enums";
 import { Badge } from "@calcom/ui/components/badge";
 import { type BadgeProps } from "@calcom/ui/components/badge";
 
-import { bookingStatusToText } from "../lib/bookingStatusToText";
+import { bookingStatusToText } from "../../lib/bookingStatusToText";
 
 export function BookingStatusBadge({ bookingStatus }: { bookingStatus: BookingStatus | null }) {
   let badgeVariant: BadgeProps["variant"] = "success";
