+++ packages/features/insights/components/booking/LowestRatedMembersTable.tsx
@@ -1,16 +1,16 @@
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 import { trpc } from "@calcom/trpc";
 
-import { useInsightsBookingParameters } from "../hooks/useInsightsBookingParameters";
-import { ChartCard } from "./ChartCard";
-import { LoadingInsight } from "./LoadingInsights";
-import { TotalUserFeedbackTable } from "./TotalUserFeedbackTable";
+import { useInsightsBookingParameters } from "../../hooks/useInsightsBookingParameters";
+import { ChartCard } from "../ChartCard";
+import { LoadingInsight } from "../LoadingInsights";
+import { TotalUserFeedbackTable } from "../TotalUserFeedbackTable";
 
-export const HighestRatedMembersTable = () => {
+export const LowestRatedMembersTable = () => {
   const { t } = useLocale();
   const insightsBookingParams = useInsightsBookingParameters();
 
-  const { data, isSuccess, isPending } = trpc.viewer.insights.membersWithHighestRatings.useQuery(
+  const { data, isSuccess, isPending } = trpc.viewer.insights.membersWithLowestRatings.useQuery(
     insightsBookingParams,
     {
       staleTime: 180000,
@@ -26,7 +26,7 @@ export const HighestRatedMembersTable = () => {
   if (!isSuccess || !data) return null;
 
   return data && data.length > 0 ? (
-    <ChartCard title={t("highest_rated")}>
+    <ChartCard title={t("lowest_rated")}>
       <TotalUserFeedbackTable data={data} />
     </ChartCard>
   ) : (
