+++ packages/features/insights/components/booking/index.ts
@@ -1,0 +1,13 @@
+export { AverageEventDurationChart } from "./AverageEventDurationChart";
+export { BookingKPICards } from "./BookingKPICards";
+export { EventTrendsChart } from "./EventTrendsChart";
+export { HighestNoShowHostTable } from "./HighestNoShowHostTable";
+export { HighestRatedMembersTable } from "./HighestRatedMembersTable";
+export { BookingsByHourChart, BookingsByHourChartContent } from "./BookingsByHourChart";
+export { LeastBookedTeamMembersTable } from "./LeastBookedTeamMembersTable";
+export { LowestRatedMembersTable } from "./LowestRatedMembersTable";
+export { MostBookedTeamMembersTable } from "./MostBookedTeamMembersTable";
+export { MostCancelledBookingsTables } from "./MostCancelledBookingsTables";
+export { PopularEventsTable } from "./PopularEventsTable";
+export { RecentFeedbackTable } from "./RecentFeedbackTable";
+export { TimezoneBadge } from "./TimezoneBadge";
