+++ packages/features/insights/components/index.ts
@@ -1,19 +1,12 @@
-export { AverageEventDurationChart } from "./AverageEventDurationChart";
-export { BookingKPICards } from "./BookingKPICards";
-export { BookingsByHourChart } from "./BookingsByHourChart";
-
-export { EventTrendsChart } from "./EventTrendsChart";
-export { FailedBookingsByField } from "./FailedBookingsByField";
-export { HighestNoShowHostTable } from "./HighestNoShowHostTable";
-export { HighestRatedMembersTable } from "./HighestRatedMembersTable";
-export { LeastBookedTeamMembersTable } from "./LeastBookedTeamMembersTable";
-export { LowestRatedMembersTable } from "./LowestRatedMembersTable";
-export { MostBookedTeamMembersTable } from "./MostBookedTeamMembersTable";
-export { MostCancelledBookingsTables } from "./MostCancelledBookingsTables";
-export { PopularEventsTable } from "./PopularEventsTable";
-export { RecentFeedbackTable } from "./RecentFeedbackTable";
-export { RoutedToPerPeriod } from "./RoutedToPerPeriod";
-export { RoutingFunnel } from "./RoutingFunnel";
-export { RoutingFormResponsesTable, type RoutingFormTableType } from "./RoutingFormResponsesTable";
-export { RoutingKPICards } from "./RoutingKPICards";
-export { TimezoneBadge } from "./TimezoneBadge";
+export { BookedByCell } from "./BookedByCell";
+export { BookingAtCell } from "./BookingAtCell";
+export { BookingStatusBadge } from "./BookingStatusBadge";
+export { CardInsights } from "./Card";
+export { CellWithOverflowX } from "./CellWithOverflowX";
+export { ChartCard } from "./ChartCard";
+export { FeedbackTable } from "./FeedbackTable";
+export { KPICard } from "./KPICard";
+export { LineChart } from "./LineChart";
+export { LoadingInsight } from "./LoadingInsights";
+export { ResponseValueCell } from "./ResponseValueCell";
+export { TotalUserFeedbackTable } from "./TotalUserFeedbackTable";
