+++ packages/features/insights/components/routing/RoutingFunnelSkeleton.tsx
